requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
djangorestframework==3.14.0
filelock==3.13.1
fsspec==2023.12.2
google-api-core==2.17.1
google-auth==2.27.0
google-cloud-language==2.13.1
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.37.1
tweepy==4.14.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
